# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  furniture-commission_internal_service.CreateRoleRequest:
    properties:
      description:
        example: 系统管理员角色
        type: string
      permission_ids:
        items:
          type: integer
        type: array
      role_code:
        example: admin
        type: string
      role_name:
        example: 管理员
        type: string
      status:
        type: integer
    required:
    - role_code
    - role_name
    type: object
...
  furniture-commission_internal_service.CreateStoreRequest:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      status:
        type: integer
      store_code:
        type: string
      store_name:
        type: string
    required:
    - address
    - store_code
    - store_name
    type: object
...
  furniture-commission_internal_service.UpdateRoleRequest:
    properties:
      description:
        example: 系统管理员角色
        type: string
      permission_ids:
        items:
          type: integer
        type: array
      role_code:
        example: admin
        type: string
      role_name:
        example: 管理员
        type: string
      status:
        type: integer
    type: object
  furniture-commission_internal_service.UpdateStoreRequest:
    properties:
      address:
        type: string
      contact_phone:
        type: string
      status:
        type: integer
      store_name:
        type: string
    type: object
...
  internal_handler.AssignPermissionsRequest:
    properties:
      permission_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  internal_handler.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
      timestamp:
        example: 1715000000
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: 家具提成管理系统后端 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 家具提成管理系统 API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: 获取所有角色列表
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新角色，需要管理员权限
      parameters:
      - description: 创建角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据角色ID删除角色，需要管理员权限
      parameters:
      - description: 角色ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详细信息
      parameters:
      - description: 角色ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 根据角色ID更新角色信息，需要管理员权限
      parameters:
      - description: 角色ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新角色请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - 角色管理
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 根据角色ID获取该角色拥有的权限列表
      parameters:
      - description: 角色ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取角色权限
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为角色分配权限，需要管理员权限
      parameters:
      - description: 角色ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 分配权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 分配权限
      tags:
      - 角色管理
...
  /stores:
    get:
      consumes:
      - application/json
      description: 获取所有门店信息列表
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取门店列表
      tags:
      - 门店管理
    post:
      consumes:
      - application/json
      description: 创建新门店（管理员）
      parameters:
      - description: 门店信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建门店
      tags:
      - 门店管理
  /stores/{id}:
    delete:
      consumes:
      - application/json
      description: 根据门店ID删除门店（管理员）
      parameters:
      - description: 门店ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除门店
      tags:
      - 门店管理
    put:
      consumes:
      - application/json
      description: 根据门店ID更新门店信息（管理员）
      parameters:
      - description: 门店ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新门店信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/furniture-commission_internal_service.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新门店
      tags:
      - 门店管理
  /upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件，支持 jpg/png/gif/bmp/webp/svg 格式，最大 5MB
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 业务错误
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: 上传图片
      tags:
      - 文件上传
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token 认证，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"